+++ src/services/youtube-api.ts
@@ -29,10 +29,7 @@ export default class {
 
   private readonly ytsrQueue: PQueue;
 
-  constructor(
-  @inject(TYPES.ThirdParty) thirdParty: ThirdParty,
-    @inject(TYPES.Config) config: Config,
-    @inject(TYPES.KeyValueCache) cache: KeyValueCacheProvider) {
+  constructor(@inject(TYPES.ThirdParty) thirdParty: ThirdParty, @inject(TYPES.Config) config: Config, @inject(TYPES.KeyValueCache) cache: KeyValueCacheProvider) {
     this.youtube = thirdParty.youtube;
     this.youtubeKey = config.YOUTUBE_API_KEY;
     this.cache = cache;